feat: replace wolfSSL with opssl TLS library
Complete migration of eggdrop's TLS backend from wolfSSL (OpenSSL-compat
API) to opssl, a custom TLS 1.2/1.3 library. This eliminates the wolfSSL
dependency and all OpenSSL compatibility shims.

Core changes:
- tls.c: rewritten for opssl_ctx/opssl_conn API
- net.c, dcc.c, botcmd.c, io_thread.c: SSL type → opssl_conn_t
- eggdrop.h, egg_tls.h: wolfSSL shim removed

Module changes:
- server/sasl.c: complete rewrite — ECDSA, SCRAM, X25519 via opssl crypto
- pbkdf2.c + tclpbkdf2.c: EVP/PKCS5 → opssl_pbkdf2/opssl_hmac
- webui.c: SHA1 → opssl_sha1_*, SSL_read → opssl_read
- dns/res.c: SSL_read/SSL_get_error → opssl_read/OPSSL_WANT_*
- md5.h: guarded extern decls for MAKING_MODS compatibility

Build system:
- Added opssl as meson subproject with full crypto + TLS
- Added pbkdf2 to tls_dep module list
- Fixed test build: added libop_inc, HAVE_CONFIG_H, libop_dep

libop additions:
- op_cidr_parse_str/op_cidr_parse_addr in cidr_tbl.c (fixes channels.so)

Tests: all 7 tests pass clean under UBSan.

+++ libop/include/op_ratelimit.h
@@ -43,6 +43,7 @@
 #define LIBOP_RATELIMIT_H
 
 #include <stdint.h>
+#include <sys/time.h>
 
 /* ---- type ---------------------------------------------------------------- */
 
@@ -87,7 +88,7 @@ op_ratelimit_init(op_ratelimit_t *rl, unsigned capacity, unsigned rate_per_sec,
                   uint64_t now_usec)
 {
     rl->rate_usec   = (rate_per_sec > 0) ? 1000000u / rate_per_sec : UINT64_MAX;
-    rl->capacity    = (uint32_t)((uint64_t)capacity * OP_RATELIMIT_SCALE);
+    rl->capacity    = capacity    * OP_RATELIMIT_SCALE;
     rl->tokens      = rl->capacity;   /* start full */
     rl->last_refill = now_usec;
 }
@@ -149,11 +150,11 @@ op_ratelimit_check_n(op_ratelimit_t *rl, uint64_t now_usec, unsigned n)
         rl->tokens = (total > rl->capacity) ? rl->capacity : (uint32_t)total;
     }
 
-    uint64_t cost = (uint64_t)n * OP_RATELIMIT_SCALE;
+    uint32_t cost = n * OP_RATELIMIT_SCALE;
     if (rl->tokens < cost)
         return false;
 
-    rl->tokens -= (uint32_t)cost;
+    rl->tokens -= cost;
     return true;
 }
